requirements.txt
selenium==4.15.0
requests==2.31.0
pillow==9.4.0
googletrans==4.0.0-rc1
webdriver-manager==3.8.6
